pkg/yoke: refactor move all takeoff command logic into commander.Takeoff
